fix(toast): create the hook for the useToast
