+++ app/(main)/evaluations/page.tsx
@@ -16,7 +16,8 @@ import PageHeader from "@/app/components/PageHeader";
 import { useAuth } from "@/app/lib/context/AuthContext";
 import { useApp } from "@/app/lib/context/AppContext";
 import { FeatureGateModal, LoginModal } from "@/app/components/auth";
-import { Loader, TabNavigation, useToast } from "@/app/components/ui";
+import { Loader, TabNavigation } from "@/app/components/ui";
+import { useToast } from "@/app/hooks/useToast";
 import DatasetsTab from "@/app/components/evaluations/DatasetsTab";
 import EvaluationsTab from "@/app/components/evaluations/EvaluationsTab";
 import { Tab } from "@/app/lib/types/evaluation";
