+++ app/components/evaluations/DatasetsTab.tsx
@@ -5,7 +5,8 @@ import { Dataset, ViewDatasetModalData } from "@/app/lib/types/dataset";
 import { useAuth } from "@/app/lib/context/AuthContext";
 import { apiFetch } from "@/app/lib/apiClient";
 import { DatabaseIcon, PlusIcon } from "@/app/components/icons";
-import { Button, Modal, useToast } from "@/app/components/ui";
+import { Button, Modal } from "@/app/components/ui";
+import { useToast } from "@/app/hooks/useToast";
 import { DatasetListSkeleton } from "@/app/components";
 import DatasetCard from "./DatasetCard";
 import CreateDatasetForm from "./CreateDatasetForm";
