+++ app/hooks/useToast.ts
@@ -1,1 +1,12 @@
-export { useToast } from "@/app/components/ui";
+"use client";
+
+import { useContext } from "react";
+import { ToastContext } from "@/app/components/ui/Toast";
+
+export function useToast() {
+  const context = useContext(ToastContext);
+  if (!context) {
+    throw new Error("useToast must be used within ToastProvider");
+  }
+  return context;
+}
